+++ src/leapfrogai_api/typedef/__init__.py
@@ -1,0 +1,1 @@
+from .model_types import Model as Model
