--- src/leapfrogai_api/types/__init__.py
@@ -1,1 +1,0 @@
-from model_types import Model as Model
